+++ src/types/spend.rs
@@ -85,6 +85,7 @@ mod tests {
             coin_spends: vec![],
             required_signatures: vec![],
             summary: TransactionSummary {
+                received: vec![],
                 outputs: vec![],
                 fee: Amount(0),
             },